sched: Implement crucial task placement strategy
Tasks that are part of the crucial schedtune group will attempt to be
placed on the idle CPU with the largest original capacity. Crucial tasks
will have the exclusion attention of the given idle CPU.

Crucial placement is not a guarenteed placement path. If an ideal CPU is
not identified, the scheduler will take the slow path in identifying a
target CPU.

It is worth noting that that `crucial` task placement strategy takes
precedence over `prefer_idle` placement. However, if a crucial CPU is
not identified (for example, if no CPUs are idle), the prefer_idle path
can still be taken.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>